scripts/requirements.txt
certifi==2023.11.17
hubspot-api-client==8.2.0
python-dateutil==2.8.2
six==1.16.0
urllib3==2.1.0
